trashweb/requirements.txt
django
djangorestframework
ultralytics
qrcode
djongo
psycopg2